requirements.txt
bs4
requests
typing
curly
biopython
pylibcurl
